+++ CHANGELOG.md
@@ -17,6 +17,10 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Update our national statistics from scorecard project
 - Consolidate how input elements are updated
 - Fix input errors if attempting to borrow more than total cost of attendance
+- Add form-level error for bad program ID
+- Update national stats to deliver generic stats at /api/national-stats/
+- simplify warnings for ID problems
+- Updated the budget section to better message lack of salary data
 
 ## 2.1.5
 - Show recalculation updates on mobile screens